Develop

+++ src/services/ActivitiesOfTheDay/DeleteMyActivityService.ts
@@ -10,6 +10,16 @@ class DeleteMyActivityService{
   async execute({ user, activity }: DeleteMyActivityServiceProps){
     const { activitiesOfTheDayRepository } = handleGetRepositories()
 
+    const activityExists = await activitiesOfTheDayRepository
+    .createQueryBuilder()
+    .where("activity = :activity", { activity })
+    .andWhere("destined_to = :user", { user })
+    .getOne()
+
+    if(!activityExists){
+      throw new Error("Activity not found status:400")
+    }
+
     await activitiesOfTheDayRepository
     .createQueryBuilder()
     .delete()
